+++ app-sh/CHANGELOG.md
@@ -1,5 +1,13 @@
 # app-sh
 
+## 1.3.5
+
+### Patch Changes
+
+- Made AppSh config optional
+
+  Now will create an AppSh for the plugin if one is not provided
+
 ## 1.3.4
 
 ### Patch Changes