+++ plugins/o365/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @app-sh/o365
 
+## 1.0.1
+
+### Patch Changes
+
+- Made appSh an optional config parameter
+- Updated dependencies
+  - app-sh@1.3.5
+
 ## 1.0.0
 
 ### Major Changes